AppVeyor is a knucklehead
AppVeyor uses the tag as the name of the "branch" and builds the tag
disconnected from the tag on the same commit, causing two builds to
happen when a commit is tagged.

This is braindead behavior, but means we can only deploy on tags and
have no way to ensure that the tag is on the master branch.

:-(